Create models for the Exploring Mars problem
Test-drive implementation of 'Map' and 'Probe' as the main
abstractions for the [Exploring Mars][1] problem.

Create auxiliary abstractions 'Point', 'Direction' and 'Command'
to make the code clearer.

Make initial implementation with immutable probes: method 'execute'
generates a new probe on each command (in a FP fashion) instead of
changing the internal state (in an OOP fashion).

Make probes stay put (aka don't move) if they try to go out of the
limits of the map. It is still necessary to handle probes trying to
access the same point at the same time (which may require the map
to track probes' current position).

[1]: https://gist.github.com/elo7-developer/1a40c96a5d062b69f02c

Signed-off-by: Renato Cordeiro Ferreira <renato.cferreira@hotmail.com>