Support serialization of maps and probes in the database
Create schema 'MarsDb' to add tables for maps and probes.

Add 'id' field in 'Map' and 'Probe' to identify the entities in the
database. Add 'idMap' field in 'Probe' to identify the map where its
inserted. Make all IDs to be UUIDs, which will be mapped to the
corresponding type ad PostgreSQL.

Change 'Probe' to inline 'Point' so this value object can be
represented with simple values that can be mapped in the database.
Add 'position' getter to keep compability with unit tests.

Add 'StatefulOneToMany' and 'StatefulManyToOne' in 'Map' and 'Probe'
respectively to make a 1-N relationship between them. Add a mapping
in the 'MarsDb' schema to create a foreign key in the probes table.

Drop the test database and regenerate the schema on each test to
guarantee the integration tests can work properly. Create a Scala
factory function 'apply' for 'Map' and 'Probe' to create new maps
and probes and store them in the database. Generate the entities
via the factory functions to store them in the database and check
their integration.

Signed-off-by: Renato Cordeiro Ferreira <renato.cferreira@hotmail.com>