Implement RESTful API to manipulate maps and probes
Implement CRUD routes in 'MapsController' and 'ProbesController'.
Create tests for the happy path and verify their exceptional behavior.

Add RPC-like routes '/maps/:mapId/probes/:probeId/execute' and
'/probes/:probeId/execute' to execute a list of commands that
will move a probe around its map.

Add special route '/database' to manage the development database so that
it is possible to create its schema with a POST and delete it with a
DELETE in the API.

Signed-off-by: Renato Cordeiro Ferreira <renato.cferreira@hotmail.com>